pyproject.toml
[project]
name = "gittask-cli"
version = "0.1.1"
description = "Git-Asana CLI & Time Tracker"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer[all]>=0.9.0",
    "questionary>=2.0.0",
    "gitpython>=3.1.0",
    "asana>=5.0.0",
    "tinydb>=4.8.0",
    "keyring>=24.0.0",
    "python-dateutil>=2.8.2",
    "thefuzz>=0.20.0",
    "PyGithub",
    "textual>=0.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-mock",
    "pytest-asyncio",
]

[project.scripts]
gittask = "gittask.main:app"
gt = "gittask.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gittask"]


[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
markers = [
    "asyncio: mark test as async",
]
